minor bug fix

+++ bernoulli/src/mc.c
@@ -25,7 +25,7 @@ void randomLabel(int * indAll, int casCount, int allCount) {
 	return;
 }
 
-int * monteCarlo(double * x, double * y, int * locEnding, int locCount, int casCount, int allCount, int wSize, int wCount, bool highLow, double * clusterLL, bool * highCluster, int nClusters, int nSim) {
+int * monteCarlo(double * x, double * y, int * locEnding, int locCount, int casCount, int allCount, double wSize, int wCount, bool highLow, double * clusterLL, bool * highCluster, int nClusters, int nSim) {
 
 	int * nExtreme;
